+++ wallets/okx/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @rosen-ui/okx-wallet
 
+## 2.2.3
+
+### Patch Changes
+
+- Update dependencies
+  - @rosen-bridge/tokens@6.0.2
+  - @rosen-network/bitcoin@2.4.6
+  - @rosen-ui/wallet-api@3.2.1
+
 ## 2.2.2
 
 ### Patch Changes